Added not-converted tests for a better structure
